feat: move oneMinus from Percent to ratio.
I'll need this in addition to addRatios and multiplyRatios.
