+++ packages/zoe/test/unitTests/contractSupport/test-percentSupport.js
@@ -4,11 +4,10 @@ import { test } from '@agoric/zoe/tools/prepare-test-env-ava';
 
 import { makeIssuerKit, amountMath } from '@agoric/ertp';
 
-import { multiplyBy, makeRatioFromAmounts } from '../../../src/contractSupport';
+import { multiplyBy } from '../../../src/contractSupport';
 import {
   make100Percent,
   make0Percent,
-  oneMinus,
 } from '../../../src/contracts/callSpread/percent';
 
 // duplicated from test-ratio, but should go away with the amount refactoring
@@ -53,28 +52,4 @@ test('ratio - NONE', t => {
     multiplyBy(moe(100000), make0Percent(brand)),
     brand,
   );
-});
-
-test('ratio - complement', t => {
-  const { brand } = makeIssuerKit('moe');
-  const moe = value => amountMath.make(value, brand);
-
-  const oneThird = makeRatioFromAmounts(moe(1), moe(3));
-  const twoThirds = oneMinus(oneThird);
-
-  amountsEqual(t, multiplyBy(moe(100000), oneThird), moe(33333), brand);
-  amountsEqual(t, multiplyBy(moe(100000), twoThirds), moe(66666), brand);
-
-  t.throws(() => oneMinus(moe(3)), {
-    message:
-      // Should be able to use more informative error once SES double
-      // disclosure bug is fixed. See
-      // https://github.com/endojs/endo/pull/640
-      //
-      // /Parameter must be a Ratio record, but .* has "brand"/,
-      /Parameter must be a Ratio record, but .* has .*/,
-  });
-  t.throws(() => oneMinus(makeRatioFromAmounts(moe(30), moe(20))), {
-    message: /Parameter must be less than or equal to 1: .*/,
-  });
 });
